Rename stub rules
