+++ crates/ruff/src/codes.rs
@@ -485,8 +485,8 @@ pub fn code_to_rule(linter: Linter, code: &str) -> Option<Rule> {
         (Flake8Pyi, "001") => Rule::PrefixTypeParams,
         (Flake8Pyi, "007") => Rule::UnrecognizedPlatformCheck,
         (Flake8Pyi, "008") => Rule::UnrecognizedPlatformName,
-        (Flake8Pyi, "009") => Rule::PreferEllipsisOverPass,
-        (Flake8Pyi, "010") => Rule::PreferOnlyEllipsis,
+        (Flake8Pyi, "009") => Rule::PassStatementStubBody,
+        (Flake8Pyi, "010") => Rule::NonEmptyStubBody,
         (Flake8Pyi, "021") => Rule::DocstringInStub,
 
         // flake8-pytest-style
