+++ crates/ruff/src/rules/flake8_pyi/mod.rs
@@ -19,10 +19,10 @@ mod tests {
     #[test_case(Rule::UnrecognizedPlatformCheck, Path::new("PYI007.py"))]
     #[test_case(Rule::UnrecognizedPlatformName, Path::new("PYI008.pyi"))]
     #[test_case(Rule::UnrecognizedPlatformName, Path::new("PYI008.py"))]
-    #[test_case(Rule::PreferOnlyEllipsis, Path::new("PYI010.py"))]
-    #[test_case(Rule::PreferOnlyEllipsis, Path::new("PYI010.pyi"))]
-    #[test_case(Rule::PreferEllipsisOverPass, Path::new("PYI009.py"))]
-    #[test_case(Rule::PreferEllipsisOverPass, Path::new("PYI009.pyi"))]
+    #[test_case(Rule::NonEmptyStubBody, Path::new("PYI010.py"))]
+    #[test_case(Rule::NonEmptyStubBody, Path::new("PYI010.pyi"))]
+    #[test_case(Rule::PassStatementStubBody, Path::new("PYI009.py"))]
+    #[test_case(Rule::PassStatementStubBody, Path::new("PYI009.pyi"))]
     #[test_case(Rule::DocstringInStub, Path::new("PYI021.py"))]
     #[test_case(Rule::DocstringInStub, Path::new("PYI021.pyi"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
