+++ crates/ruff/src/rules/flake8_pyi/rules/mod.rs
@@ -1,13 +1,13 @@
 pub use docstring_in_stubs::{docstring_in_stubs, DocstringInStub};
-pub use prefer_ellipsis_over_pass::{prefer_ellipsis_over_pass, PreferEllipsisOverPass};
-pub use prefer_only_ellipsis::{prefer_only_ellipsis, PreferOnlyEllipsis};
+pub use non_empty_stub_body::{non_empty_stub_body, NonEmptyStubBody};
+pub use pass_statement_stub_body::{pass_statement_stub_body, PassStatementStubBody};
 pub use prefix_type_params::{prefix_type_params, PrefixTypeParams};
 pub use unrecognized_platform::{
     unrecognized_platform, UnrecognizedPlatformCheck, UnrecognizedPlatformName,
 };
 
 mod docstring_in_stubs;
-mod prefer_ellipsis_over_pass;
-mod prefer_only_ellipsis;
+mod non_empty_stub_body;
+mod pass_statement_stub_body;
 mod prefix_type_params;
 mod unrecognized_platform;
