+++ crates/ruff/src/rules/flake8_pyi/rules/pass_statement_stub_body.rs
@@ -1,5 +1,6 @@
+use rustpython_parser::ast::{Stmt, StmtKind};
+
 use ruff_macros::{define_violation, derive_message_formats};
-use rustpython_parser::ast::{Located, StmtKind};
 
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
@@ -7,23 +8,23 @@ use crate::registry::Diagnostic;
 use crate::violation::Violation;
 
 define_violation!(
-    pub struct PreferEllipsisOverPass;
+    pub struct PassStatementStubBody;
 );
-impl Violation for PreferEllipsisOverPass {
+impl Violation for PassStatementStubBody {
     #[derive_message_formats]
     fn message(&self) -> String {
-        format!("Empty body should contain '...', not 'pass'")
+        format!("Empty body should contain `...`, not `pass`")
     }
 }
 
 /// PYI009
-pub fn prefer_ellipsis_over_pass(checker: &mut Checker, body: &Vec<Located<StmtKind>>) {
+pub fn pass_statement_stub_body(checker: &mut Checker, body: &[Stmt]) {
     if body.len() != 1 {
         return;
     }
-    if body[0].node == StmtKind::Pass {
+    if matches!(body[0].node, StmtKind::Pass) {
         checker.diagnostics.push(Diagnostic::new(
-            PreferEllipsisOverPass,
+            PassStatementStubBody,
             Range::from_located(&body[0]),
         ));
     }
